All user namespaces should be listed in owned
We should always list all organizations and namespaces a user is part of
in "owned" folder listing. Current behavior was that as soon as one
namespace created a notebook we'd only list that on. The change makes it
so we always build the base list of namespaces in the owned category.