Pull pure path manipulation into paths.py.
This creates the paths module and pulls out all the path work where
we're just twiddling with path strings into that module, to separate it
from the rest of tiledbcontents. It also removes all use of `os.sep`,
because per the Jupyter documentation, all paths in the Jupyter API are
passed with slash separators:

https://jupyter-notebook.readthedocs.io/en/stable/extending/contents.html#apipaths

Functions in the paths module have docstrings that can be run through
doctest.