Use a long-ago date when filling in missing datetimes. (#64)
Previously, we learned that Jupyter REALLY REALLY WANTS to know the
creation and modification dates of every single thing it touches
and is NOT satisfied with being told that said information simply
does not exist. So we decided to just tell it "OK, the time is now.
Are you happy?" And initially, it did seem happy! We could successfully
browse and open and all the file operations we love to do so much,
so far as we could tell. But then, we found out that there were,
in fact, Problems.

Jupyter (sensibly) uses the last-modification date to check for edit
races, so that if you make a change but the file changes out under you
(i.e., its edit date changes since the time that you opened it),
it gives you a warning that you may be overwriting others' changes.
Given that we were always telling Jupyter that the last-modified time
was Right Now, Jupyter would always think that the file had been changed
and warn the user that they might be overwriting existing changes.

So now tell Jupyter a different lie: if we don't know what time
something happened at, we just say it was an arbitrary time
in the past. (That arbitrary time happens to be the timestamp of
f7f1165d96766ae311183d7c2912b1eff2a5451b, the first commit
to the TileDB core repository.)