Upgrade dependencies for TileDB
We need TileDB-Py >= 0.15.5 because a new type TILEDB_BLOB is introduced
in this release and we will use it for the notebook schema.

Also 0.16.1 uses the latest tiledb core 2.10.1

TileDB-Cloud version is not crucial. We use simple operations like
  tiledb.cloud.array.info
  tiledb.cloud.array.update_info
  tiledb.cloud.client.user_profile
  tiledb.cloud.client.organization
  tiledb.cloud.client.list_arrays
  tiledb.cloud.array.delete_array
  tiledb.cloud.notebook.rename_notebook
which are not changed frequently. But since we use this module only
internally we can update to the latest release.

Ref https://app.shortcut.com/tiledb-inc/story/18832/update-jupyter-contents-api-to-latest-tiledb-py