Simplify caching layer and make it async-aware.
This updates the caching layer to only cache metadata rather than also
holding onto array contents and an opened Array. This is because,
empirically, jupyter (no longer?) does a read immediately after a write,
which was the rationale for caching notebook contents.

Instead, we only open the array as needed to read/write it, and load
and update metadata at those times. This (and some other functions)
now run in an async-aware manner, putting only the blocking I/O onto
a worker thread and managing all other data in the main event loop.