Make downloading work right (again? for the first time?) (#79)
The [`Async`]`FileContentsManager` uses a special static content handler
to handle file downloads that does not interact with the Jupyter
contents manager at all. This meant that when we subclassed it, requests
for file downloads (which work differently for requests for notebook
contents, because ???) would be intercepted before they even hit the
rest of Jupyter meaning that it was nigh impossible to figure out what
was going on because even when you did a download, no request would
appear in the server-side request logs on stderr.

After way too much digging I eventually figured out about the default
files handler class:

https://github.com/jupyter-server/jupyter_server/blob/afb1af03bb17df59ddd1d29ba6baf6be6678d0d3/jupyter_server/services/contents/filemanager.py#L80-L86

and that this is what we needed to override. And this brings us here.