Use the show granular storage paths if a path is not explicitly set
This is about the s3 uri for array creation. In production the UI
always sets it explicitly using the granular paths if specified,
otherwise it fallsback to the default s3 path.
In development the default s3 path is always used ignoring
the granular paths. This commit uses the granular paths for dev.