Remove suffix striping from guess_type
The method should not change the path. It is the caller's
responsibility to pass a properly stripped path.

Jupyter needs names with the .ipynb suffix but TileDB arrays do
not have it by convention. Our code adds the suffix
when passing paths to Jupyter and removes it when passing paths
back to cloud. If an array name already has .ipynb suffix then guess_type
did a second unneeded strip that caused an error. For example

notebook.ipynb is the cloud notebook
notebook.ipynb.ipynb is sent to Jupyter
.. do some jupyter work
notebook.ipynb is about to be sent back
notebook is what guess_type checks

notebook does not exist so we get an error because Jupyter considers it
a directory

See https://app.shortcut.com/tiledb-inc/story/25796/notebook-with-ipynb-ipynb-fails-to-load