Check if a named object exists *before* trying to create it. (#90)
TileDB Cloud used to forbid objects with duplicate names (e.g., you
couldn't have two arrays with the same name). However, because distinct
groups may contain entries with the same name, this is no longer the
case. For this reason, we need to check for an object's existence
at a given name before attempting to create a new one.