Enable using TileDB array timestamps for remote checkpoints.
This reads the fragment timestamps of writes to the array to assemble
the list of checkpoints the user can restore to.  Checkpoint IDs are
generated (and parsed) based on the timestamp in a human-readable
YYYY-MM-DD HHhMM[mSS[sSSS]] format (without showing unnecessary
precision when there is no ambiguity).

Verified working in a Jupyter Lab installation.