Rip out caching machinery and prepare for new Jupyterlab.
We ran into a problem when deleting a notebook and creating a new one
of the same name. We believe this might be related to caching. Even if
it isn't, the caching system is no longer particularly useful.

This change removes the caching machinery, and at the same time fixes
several bugs and prepares us for the next version of Jupyterlab, which
will require contents plugins to include the hash of the notebook file
when reading it (i.e., everything that needed to be done to make it work
on a freshly-installed Jupyterlab venv).