Handle getting `None` arrays back from TileDB Cloud. (#102)
When deserializing a notebook listing response from TileDB Cloud,
it is possible that the generated client code handles the zero-array
case by setting the `arrays` value to `None` rather than an empty
iterable. This would cause a `NoneType is not iterable` exception.

I'm not *certain* if this is the case, but it's a guess.